build: reorganize cli scripts for poetry format
This way, the CLI scripts are automatically installed when running
`poetry install` and are available for use as described in the README.
