+++ Makefile
@@ -5,12 +5,9 @@ dev:
 check:
 	isort -c hack/
 	isort -c scripts/
-	isort -c bin/
 	black --check hack/
 	black --check scripts/
-	black --pyi --check bin/*
 	flake8 hack/
 	flake8 scripts/
-	flake8 bin/
 
 .PHONY: dev check
